requirements.txt
greenlet==1.1.3
kafka-python==2.0.2
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
